% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cccounter.R
\name{cccounter}
\alias{cccounter}
\title{Generate Counterfactual Scenarios Based on Temperature Changes}
\usage{
cccounter(m, beta, sigma, N, psihat, vec.new.trend, H = 100)
}
\arguments{
\item{m}{An integer representing the number of past observations used in the calculation.}

\item{beta}{A numeric value representing the base parameter for temperature changes.}

\item{sigma}{A numeric value representing the standard deviation of temperature changes.}

\item{N}{An integer representing the number of countries or observations.}

\item{psihat}{A numeric vector of psi coefficients.}

\item{H}{An integer specifying the number of time periods for the counterfactual scenarios. Default is 100.}

\item{delta.trend}{A scalar representing changes in temperature trends for each period.}
}
\value{
A list containing:
\item{Delta}{A matrix of changes in temperatures for the counterfactual scenarios.}
\item{mu}{A matrix of mean expected temperature changes.}
\item{omega}{A matrix of standard deviations for temperature changes.}
\item{g0}{A numeric value representing the closed form for the expectation of the absolute change in temperatures.}
}
\description{
This function generates counterfactual scenarios given a scenario of changes
in temperatures for all countries, based on specified trends and volatilities.
}
