% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.coef.se.dm.R
\name{ec.coef.se.dm}
\alias{ec.coef.se.dm}
\title{Calculate long-run model coefficients and their Standard Errors from Error Correction Model using the Delta Method}
\usage{
ec.coef.se.dm(fitted.model, numreg = 4)
}
\arguments{
\item{fitted.model}{A fitted model object that contains the coefficients and
variance-covariance matrix (vcov) of the model.}

\item{numreg}{The number of regressors for which we wish to compute the long-run impacts and their .
standard errors. Default to 2.}
}
\value{
A list containing:
\item{lr.coef}{A vector of log-ratio coefficients.}
\item{lr.coef.se}{A vector of standard errors for the log-ratio coefficients.}
}
\description{
This function calculates the long run coefficients of an ARDL and their standard errors
using the delta method based on its error correction form. It takes a fitted model
as input and computes the transformed coefficients as well as their standard errors.
}
\examples{
# Assuming `model` is a fitted model object
result <- lr.coef.se.dm(model)

}
