% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpj.fe.R
\name{hpj.fe}
\alias{hpj.fe}
\title{hpj.fe}
\usage{
hpj.fe(formula, dataset, effect, model = "within", compute.std.err = TRUE)
}
\arguments{
\item{formula}{A formula specifying the model to be estimated.}

\item{dataset}{A pdata frame that contains the data for the analysis.}

\item{effect}{A character string specifying the type of effect ("individual" or "time").}

\item{model}{A character string specifying the model type ("within", "random", "pooling").
Default is "within".}

\item{compute.std.err}{A logical indicating whether to compute standard errors. Default is TRUE.}
}
\value{
A list containing coefficients, standard errors, fitted values, residuals,
variance-covariance matrix, and the dataset used.
}
\description{
A function to perform half-panel jacknife estimation of ARDL panel data models
using the plm package.
}
