% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpj.het.R
\name{hpj.het}
\alias{hpj.het}
\title{hpj.het}
\usage{
hpj.het(formula, dataset)
}
\arguments{
\item{formula}{A formula specifying the model to be estimated.}

\item{dataset}{A pdata frame that contains the data for the analysis.}
}
\value{
A list containing the half-panel jackknife heterogeneous estimates and the estimates
from the full and each of the half panels.
}
\description{
A function to perform half-panel jacknife estimation of ARDL heterogenous panel data models
using the plm package.
}
