% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psicoef.b.R
\name{psicoef.b}
\alias{psicoef.b}
\title{Generate Bootstrap for the Psi Coefficients}
\usage{
psicoef.b(p.dataset, formula.ardl, thetaest, residuals, n.years = 101, B = 99)
}
\arguments{
\item{p.dataset}{A pdata.frame containing the dataset for the analysis.}

\item{formula.ardl}{A formula specifying the ARDL model.}

\item{thetaest}{A vector of estimated coefficients from the model.}

\item{residuals}{A vector of residuals from the model.}

\item{n.years}{An integer specifying the number of years for which to generate coefficients. Default is 101.}

\item{B}{An integer specifying the number of bootstrap repetitions. Default is 99.}
}
\value{
A list containing:
\item{psihat}{A matrix of psi coefficients for each bootstrap sample.}
\item{thetahat}{A matrix of estimated coefficients for each bootstrap sample.}
}
\description{
This function performs a dynamic wild bootstrap to generate estimates for the
psi coefficients based on the provided dataset and ARDL model.
}
