% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psicoef.R
\name{psicoef}
\alias{psicoef}
\title{Generate Coefficients of the MA(∞) Representation of the ARDL Model}
\usage{
psicoef(hpj.p.1, n.years = 101)
}
\arguments{
\item{hpj.p.1}{A list containing the coefficients from the ARDL model.
It should have a named component \code{coeff} with at least 9 coefficients.}

\item{n.years}{An integer specifying the number of years for which to generate coefficients.
Default is 101.}
}
\value{
A vector of coefficients for the MA(∞) representation for the specified number of years.
}
\description{
This function computes the coefficients for the MA(∞) representation of the ARDL model
based on the input coefficients from a previous estimation.
}
