% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr.coef.se.dm.R
\name{sr.coef.se.dm}
\alias{sr.coef.se.dm}
\title{Calculate long-run model coefficients and their Standard Errors using the Delta Method}
\usage{
sr.coef.se.dm(fitted.model, p = 4, q = p + 1, numreg = 4)
}
\arguments{
\item{fitted.model}{A fitted model object that contains the coefficients and
variance-covariance matrix (vcov) of the model.}

\item{p}{Order of the auto-regressive part. Default to 4.}

\item{q}{Lags of the exogenous variables. Default to p + 1.}

\item{numreg}{The number of regressors for which we wish to compute the long-run impacts and their .
standard errors. Default to 2.}
}
\value{
A list containing:
\item{lr.coef}{A vector of log-ratio coefficients.}
\item{lr.coef.se}{A vector of standard errors for the log-ratio coefficients.}
}
\description{
This function calculates the long-run coefficients and their standard errors
using the delta method based on a fitted model object. It takes a fitted model
as input and computes the transformed coefficients as well as their standard errors.
}
\examples{
# Assuming `model` is a fitted model object
result <- sr.coef.se.dm(model)

}
