% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createlagvars.R
\name{createlagvars}
\alias{createlagvars}
\title{Create Lagged Variables}
\usage{
createlagvars(initdata, ma)
}
\arguments{
\item{initdata}{A pdata.frame containing the data for which lagged variables will be created.}
}
\value{
A pdata.frame with additional lagged variables and aggregate world GDP.
}
\description{
This function creates lagged variables for various growth and precipitation metrics,
as well as computes counterfactual variables based on temperature differences.
}
