% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DWB.R
\name{DWB}
\alias{DWB}
\title{Dynamic Wild Bootstrap Function}
\usage{
DWB(y)
}
\arguments{
\item{y}{A numeric vector of observations for which the bootstrap is to be performed.}
}
\value{
A numeric vector of bootstrapped values.
}
\description{
This function performs dynamic wild bootstrap on the input data using
the approach outlined in Gao et al. It utilizes a Bartlett kernel and
computes the necessary adjustments for the bootstrap procedure.
}
